requirements.txt
paho-mqtt==1.6.1
discord.py==2.3.2
SQLAlchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0